% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnews.R
\name{getnews}
\alias{getnews}
\title{Get all the NEWS entries that concern a particular function}
\usage{
getnews(fn, pkg = NULL, recent_first = TRUE)
}
\arguments{
\item{fn}{Function to search}

\item{pkg}{In which package is the function? Default is `NULL`, which searches
in base R.}

\item{recent_first}{Display the most recent version at the bottom of the
console? Default is `TRUE`.}
}
\value{
Prints text
}
\description{
Get all the NEWS entries that concern a particular function
}
\examples{
getnews('list2DF')
getnews("paste0")
getnews("sprintf")
getnews("relocate", "dplyr", recent_first = TRUE)
}
